django_projects/requirements.txt
Django
psycopg2-binary
python-decouple
djangorestframework
pillow
